django/ssp/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
paho-mqtt==1.6.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
